[feat] 상세화면 UI 및 기능 구현
