+++ Soulmate/Soulmate/Presentation/HomeScene/ViewModel/DetailViewModel.swift
@@ -7,7 +7,23 @@
 
 import Foundation
 
-struct DetailViewModel {
-
+final class DetailViewModel {
+    let userInfo: RegisterUserInfo
+    let distance: Int
     
+    init(userInfo: RegisterUserInfo, distance: Int) {
+        self.userInfo = userInfo
+        self.distance = distance
+    }
+    
+}
+
+extension DetailViewModel {
+    struct Input { }
+    
+    struct Output { }
+    
+    func transform(input: Input) -> Output {
+        return Output()
+    }
 }
